requirements.txt
Flask==2.2.5
Flask-Cors==3.1.1
Pillow==10.0.0
PyMuPDF==1.22.3
gunicorn==20.1.0
